+++ lib/models/income.dart
@@ -7,7 +7,7 @@ class Income {
   String name = '';
   IncomeFrequency frequency = IncomeFrequency.Monthly;
   int amount = 0;
-  int yearlyAppreciationPercentage = 0;
+  double yearlyAppreciationPercentage = 0;
   Jiffy startDate = Jiffy().startOf(Units.DAY);
   Jiffy endDate = Jiffy().startOf(Units.DAY).add(years: 100);
   bool belongsToAsset = false;
